Remove dependency on NLog
Unless NLog.dll is present beside DataProvider.dll, attempts by the F#
compiler to use DataProvider.dll will fail.

That causes problems when distributing with NuGet. If NLog.dll is included
in the package, this could silently conflict with dependencies on NLog.
If NLog.dll is not included in the package, but listed as a dependency
instead, building will fail because the F# compiler can't load NLog.dll.

As a work-around, we resort to implementing simple logging manually.