Intercept exceptions in logging
I'm not quite sure it's impossible for a type provider to be invalidated
while it's methods are stil being executed. This kind of situation could
lead to logging attempting to write to a file that's been closed. Should
that happen, it should not be allowed to crash the type provider.