Attempt to deal with parameters with generated types in generated methods
The problem is that one cannot use quotation holes, as the type of the
argument has a dynamically created type, and the syntax of holes does
not allow for using type objects, only type names.

I could not find a constructor for holes in Expr, so I used the closest
thing, a coercion expression. I don't expect it will work for options and
lists though.