Conversion of ValueType to Expr now uses procedural code
This avoids stack overflows during compilation of provided code.

It's possible the same effect can be achieved using functional code.

One benefit of this approach is to show how to create and use variables
in quotations, which might lead to an alternative to quotation holes for
injecting expressions in expressions. The problem with quotation holes is
that they require exact type matches, which I don't think works when
the received type is a generated type.