Add Expr conversion helpers to work around limitations with quotation holes
The limitation is that putting a value with type B in a hole with type A
is only allowed if A and B are the exact same type.

This means one cannot consume arguments of generated types in quotations,
because the generated type cannot be used in the syntax of quotation holes.