Generate code that relies on reflection to convert maps
That works, for now, but I wouldn't be surprised if it failed at runtime.
This also means that the generated code cannot be used on a platform that
does not support Reflection.Emit.

The alternative of using reflection at compile time is something I could
not manage to get to work. The problem seems to be that instantiating
generic types (such as Map) and methods with generated types doesn't work
when using Type.MakeGenericType, and even using
ProvidedTypeBuilder.MakeGenericType is problematic (doesn't find code in
the target assembly).